backend/app/requirements.txt
streamlit>=1.31
python-dotenv>=1.0.0
pypdf>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.1
openai>=1.0.0
numpy>=1.26
